Fix some warnings

+++ src/windows/title_editor.c
@@ -574,7 +574,7 @@ static void window_title_editor_mousedown(int widgetIndex, rct_window* w, rct_wi
 			widget--;
 			for (i = 0; i < num_items; i++) {
 				gDropdownItemsFormat[i] = 2777;
-				gDropdownItemsArgs[i] = (uint64)&gConfigTitleSequences.presets[i].name;
+				gDropdownItemsArgs[i] = (uint32)&gConfigTitleSequences.presets[i].name;
 			}
 
 			window_dropdown_show_text_custom_width(
